quote_asset_entry_amount as i64

+++ programs/clearing_house/src/state/events.rs
@@ -404,7 +404,7 @@ pub struct SettlePnlRecord {
     pub pnl: i128,
     pub base_asset_amount: i128,
     pub quote_asset_amount_after: i128,
-    pub quote_entry_amount: i128,
+    pub quote_entry_amount: i64,
     pub settle_price: i128,
 }
